+++ Platformio/platformio.ini
@@ -30,7 +30,7 @@ build_flags =
 	-D DISABLE_ALL_LIBRARY_WARNINGS=1 ; for TFT_eSPI
 	-D USE_NIMBLE=1 ; for BLE Keyboard
 	-D CORE_DEBUG_LEVEL=ARDUHAL_LOG_LEVEL_DEBUG
-	-D LV_CONF_PATH=../../../../src/gui_general/lv_conf.h
+	-D LV_CONF_PATH=../../../../src/gui_general_and_keys/lv_conf.h
 	; The following lines replace the TFT_eSPI User_setup.h-file
 	-D USER_SETUP_LOADED=1
 	-D ILI9341_DRIVER=1
