+++ Platformio/src/device_smarthome/device_smarthome_assets.c
