+++ Platformio/src/gui_general_and_keys/gui_base_assets.c
